new lingos and 2byte cmd support

+++ lingo-general/handler.go
@@ -32,11 +32,11 @@ type DeviceGeneral interface {
 }
 
 func ackSuccess(req ipod.Packet) ACK {
-	return ACK{Status: ACKStatusSuccess, CmdID: req.ID.CmdID()}
+	return ACK{Status: ACKStatusSuccess, CmdID: uint8(req.ID.CmdID())}
 }
 
 func ackPending(req ipod.Packet, maxWait uint32) ACKPending {
-	return ACKPending{Status: ACKStatusPending, CmdID: req.ID.CmdID(), MaxWait: maxWait}
+	return ACKPending{Status: ACKStatusPending, CmdID: uint8(req.ID.CmdID()), MaxWait: maxWait}
 }
 
 func ackFIDTokens(tokens SetFIDTokenValues) RetFIDTokenValueACKs {
